% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-ExpressionResultSet.R
\name{consensus}
\alias{consensus}
\alias{consensus,ExpressionResultSet-method}
\title{Get consensus results}
\usage{
consensus(object)

\S4method{consensus}{ExpressionResultSet}(object)
}
\arguments{
\item{object}{An ExpressionResultSet object.}
}
\value{
A list.
}
\description{
Get consensus results
}
\examples{
obj <- createExpressionResultSet(
  results = list(edgeR = data.frame(gene = c("g1","g2"), logFC = c(1,-1))),
  methodNames = "edgeR",
  consensus = list(g1 = TRUE)
)
consensus(obj)
}
