% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consexpressionR-package.R
\docType{package}
\name{consexpressionR-package}
\alias{consexpressionR-package}
\title{consexpressionR: consexpressionR: Consensus of expression analysis with RNA-Seq data.}
\description{
Consensus among five Diferential Expressed Genes (DEGs) Analysis for RNA-Seq data. Identification methods and, with a consensus result. Various methods DEGs Analysys in one tool.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://costasilvati.github.io/consexpressionR/}
}

}
\keyword{internal}
