% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consexpressionR-package.R
\docType{package}
\name{consexpressionR}
\alias{consexpressionR-package}
\alias{consexpressionR}
\title{consexpressionR: Differential Expression Analysis using consensus of multiple methods}
\description{
Comprehensive workflow for differential expression (DE) analysis using multiple
methods (e.g., DESeq2, edgeR, limma, EBSeq, NOISeq, SAMSeq, KnowSeq) and
consensus strategies to derive robust DE gene lists.
}
\examples{
obj <- createExpressionResultSet(
  results = list(edgeR = data.frame(gene = "g1", logFC = 1)),
  methodNames = "edgeR"
)
obj
}
\seealso{
Useful links:
\itemize{
  \item \url{https://costasilvati.github.io/consexpressionR/}
  \item Report bugs at \url{https://github.com/costasilvati/consexpressionR/issues}
}

}
\author{
\strong{Maintainer}: Juliana Costa-Silva \email{julianacostasilvati@gmail.com} (\href{https://orcid.org/0000-0002-5830-8548}{ORCID}) [funder]

Authors:
\itemize{
  \item David Menotti \email{menottid@gmail.com} (\href{https://orcid.org/0000-0003-2430-2030}{ORCID})
  \item Fabricio M. Lopes \email{fabricio@utfpr.edu.br} (\href{https://orcid.org/0000-0002-8025-6422}{ORCID})
}

}
