% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createNameFileOutput.R
\name{createNameFileOutput}
\alias{createNameFileOutput}
\title{Build a file output name}
\usage{
createNameFileOutput(outDirPath = ".", execName = "exec")
}
\arguments{
\item{outDirPath}{the name of the file which the data are to be read from. Each row of the table appears as one line of the file. If it does not contain an absolute path, the file name is relative to the current working directory, getwd(). This can be a compressed file (see file). Alternatively, file can be a readable text-mode connection (which will be opened for reading if necessary, and if so closed (and hence destroyed) at the end of the function call).}

\item{execName}{name of the tool run, can be a gene expression tool}

\item{experimentName}{the name of experiment analised by consexpression2}
}
\value{
string name of path file output
}
\description{
Build a file output name
}
\examples{
nameFile <- createNameFileOutput ("../", "new_experiment", "test_exec")
}
