% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-ExpressionResultSet.R
\docType{class}
\name{ExpressionResultSet-class}
\alias{ExpressionResultSet-class}
\title{An S4 class to store differential expression results}
\description{
An S4 class to store differential expression results
}
\section{Slots}{

\describe{
\item{\code{results}}{A named list of data.frames, each corresponding to a method (e.g., edgeR, DESeq2, etc.)}

\item{\code{methods}}{A character vector of methods used}

\item{\code{parameters}}{A list of parameters used in the analysis}

\item{\code{consensus}}{A list containing consensus DEG results (e.g., from consensusList)}
}}

