% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatMapPlot.R
\name{heatMapPlot}
\alias{heatMapPlot}
\title{Makes a heatmap plot with differential expressed genes}
\usage{
heatMapPlot(
  countData,
  designExpreiment,
  pdfOutName = "heatMap.pdf",
  condition = "default",
  writePdf = FALSE
)
}
\arguments{
\item{pdfOutName}{string: path + name file, to write a pdf file with Plot (default: "heatMap.pdf")}

\item{condition}{string: name of analysed experminet (default: "condition_default")}

\item{writePdf}{logical variable: TRUE print report by each tool, FALSE print only consensus result}
}
\description{
Makes a heatmap plot with differential expressed genes
}
