% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-ExpressionResultSet.R
\name{methodNames}
\alias{methodNames}
\title{Get method names used in the analysis}
\usage{
methodNames(object)
}
\arguments{
\item{object}{An ExpressionResultSet object.}
}
\value{
A character vector.
}
\description{
Get method names used in the analysis
}
\examples{
obj <- createExpressionResultSet(
  results = list(edgeR = data.frame(gene = "g1", logFC = 1)),
  methodNames = c("edgeR")
)
methodNames(obj)
}
