% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-ExpressionResultSet.R
\name{parameters}
\alias{parameters}
\title{Get parameters used in the analysis}
\usage{
parameters(object)
}
\arguments{
\item{object}{An ExpressionResultSet object.}
}
\value{
A list.
}
\description{
Get parameters used in the analysis
}
\examples{
obj <- createExpressionResultSet(
  results = list(edgeR = data.frame(gene = "g1", logFC = 1)),
  methodNames = "edgeR",
  parameters = list(alpha = 0.01)
)
parameters(obj)
}
