% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCountFile.R
\name{readCountFile}
\alias{readCountFile}
\title{Read table count File and creates a matrix from the given content file.}
\usage{
readCountFile(tableCountPath = "data/table_count_df.csv", split = ",")
}
\arguments{
\item{tableCountPath}{the name of the file which the data are to be read from. Each row of the table appears as one line of the file. If it does not contain an absolute path, the file name is relative to the current working directory, getwd(). This can be a compressed file (see file). Alternatively, file can be a readable text-mode connection (which will be opened for reading if necessary, and if so closed (and hence destroyed) at the end of the function call).}

\item{split}{the field separator character. Values on each line of the file are separated by this character. If sep = "" (the default for read.table) the separator is ‘,’, newlines or carriage returns.}
}
\value{
content of file in matrix R format
}
\description{
Read table count File and creates a matrix from the given content file.
}
\examples{
df <- readCountFile(tableCountPath="data/table_count_df.csv",
                    split=",")
}
