% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-ExpressionResultSet.R
\name{results}
\alias{results}
\alias{results,ExpressionResultSet-method}
\title{Get result tables}
\usage{
results(object)

\S4method{results}{ExpressionResultSet}(object)
}
\arguments{
\item{object}{An ExpressionResultSet object.}
}
\value{
A named list of data.frames (or NULL elements).
}
\description{
Get result tables
}
\examples{
obj <- createExpressionResultSet(
  results = list(edgeR = data.frame(gene = "g1", logFC = 1)),
  methodNames = "edgeR"
)
results(obj)
}
