% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runConsexpression2.R
\name{runConsexpression2}
\alias{runConsexpression2}
\title{Execute app shinny version}
\usage{
runConsexpression2()
}
\value{
void
}
\description{
Execute app shinny version
}
\examples{
runConsexpression2()
}
