% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runEbseq.R
\name{runEbseq}
\alias{runEbseq}
\title{Execute EBSeq gene Expression analysis}
\usage{
runEbseq(
  countMatrix,
  designExperiment,
  fdr = 0.05,
  ppThreshold = 0.8,
  maxRound = 50,
  methodDeResults = "robust",
  groups = c("")
)
}
\arguments{
\item{countMatrix}{either a matrix of raw (read) counts.}

\item{designExperiment}{replicate and treatment by samples}

\item{fdr}{parameter used in EBTest function: fdr False Discovery Rate cutt off}

\item{ppThreshold}{posterior Probability Threshold}

\item{maxRound}{parameter used in EBTest function: Number of iterations. The default value is 50.}

\item{methodDeResults}{parameter used in GetDEResults function: "robust" or "classic". Using the "robust" option, EBSeq is more robust to genes with outliers and genes with extremely small variances. Using the "classic" option, the results will be more comparable to those obtained by using the GetPPMat() function from earlier version (<= 1.7.0) of EBSeq}

\item{groups}{text, name of samples or treatment}
}
\value{
EBSeq report in data Frame fromat
}
\description{
Execute EBSeq gene Expression analysis
}
\examples{
set.seed(42)
counts <- matrix(
  as.integer(c(
    rnbinom(200, mu = 10, size = 1),
    rnbinom(200, mu = 100, size = 1)
  )),
  nrow = 100,
  dimnames = list(
    paste0("gene", seq_len(100)),
    paste0("sample", seq_len(4))
  )
)
groups_info <- c("control", "control", "treated", "treated")
treats <- c("control", "treated")
designExperimentModel <- rep(treats, each = 2)
toolResult <- NULL
toolResult$ebseq <- runEbseq(counts, designExperimentModel, groups = treats)
}
