% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runEdger.R
\name{runEdger}
\alias{runEdger}
\title{Execute edgeR expression Analisys}
\usage{
runEdger(countMatrix, numberReplics, desingExperiment, methNorm = "TMM")
}
\arguments{
\item{countMatrix}{either a matrix of raw (read) counts.}

\item{numberReplics}{number of replicate (technical or biologcal) integer}

\item{desingExperiment}{replicate and treatment by samples}

\item{methNorm}{normalization method to be used in edgeR::calcNormFactors(), default: "TMM"}
}
\value{
edgeR report in data Frame
}
\description{
Execute edgeR expression Analisys
}
\examples{
set.seed(42)
counts <- matrix(
  as.integer(c(
    rnbinom(200, mu = 10,  size = 1), 
    rnbinom(200, mu = 100, size = 1) 
  )),
  nrow = 100,
  dimnames = list(paste0("gene", seq_len(100)),
                    paste0("sample", seq_len(4)))
)
treats <- c("control", "treated")
numberReplicsModel <- 2
designExperimentModel <- rep(treats, each = numberReplicsModel)
toolResult <- NULL
toolResult$edger <- runEdger(countMatrix = counts, numberReplics = numberReplicsModel,
                               desingExperiment = designExperimentModel)
}
