% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLimma.R
\name{runLimma}
\alias{runLimma}
\title{Execute edgeR expression Analisys}
\usage{
runLimma(
  countMatrix,
  numberReplics,
  designExperiment,
  limmaOutPath,
  methodNorm = "TMM",
  methodAdjPvalue = "BH",
  numberTopTable = 1e+06
)
}
\arguments{
\item{countMatrix}{either a matrix of raw (read) counts.}

\item{numberReplics}{number of replicate (technical or biologcal) integer}

\item{designExperiment}{replicate and treatment by samples}

\item{limmaOutPath}{path to write limma report in csv file format}

\item{methodNorm}{normalization method to be used}

\item{methodAdjPvalue}{correction method, a character string. Can be abbreviated.}

\item{numberTopTable}{maximum number of genes to list}
}
\value{
limma report in data Frame fromat
}
\description{
Execute edgeR expression Analisys
}
\examples{
limmaResult<-runLimma(countMatrix, numberReplics, designExperiment, createNameFileOutput(outDirPath,experimentName,execName='limma'))
}
