% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSamSeq.R
\name{runSamSeq}
\alias{runSamSeq}
\title{Execute SAMSeq gene Expression analysis to count data}
\usage{
runSamSeq(
  countMatrix,
  designExperiment,
  respType = "Two class unpaired",
  numberPermutations = 100
)
}
\arguments{
\item{countMatrix}{either a matrix of raw (read) counts.}

\item{respType}{Problem type: "Quantitative" for a continuous parameter; "Two class unpaired" for two classes with unpaired observations; "Survival" for censored survival outcome; "Multiclass": more than 2 groups; "Two class paired" for two classes with paired observations.}

\item{desingExperiment}{replicate and treatment by samples}

\item{numerPermutations}{Number of permutations used to estimate false discovery rates}
}
\value{
SAMSeq report in data Frame
}
\description{
Execute SAMSeq gene Expression analysis to count data
}
