% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-ExpressionResultSet.R
\name{summary,ExpressionResultSet-method}
\alias{summary,ExpressionResultSet-method}
\title{Summary method for ExpressionResultSet}
\usage{
\S4method{summary}{ExpressionResultSet}(object)
}
\arguments{
\item{object}{An ExpressionResultSet object}
}
\value{
Invisibly returns \code{NULL}. This method is used for its
side effect of printing a summary to the console.
}
\description{
Summary method for ExpressionResultSet
}
