% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_count_df.R
\docType{data}
\name{table_count_df}
\alias{table_count_df}
\title{Meu objeto incrível}
\format{
An object of class \code{character} of length 1.
}
\usage{
table_count_df
}
\value{
data.frame with 31 lines (genes) and 6 columns (treatments BM and JJ)
}
\description{
Dataset contains GSE95077 NCBI GEO dataset
}
\examples{
table_count_df


}
\keyword{table_count_df}
