% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upSetPlotTools.R
\name{upSetPlotTools}
\alias{upSetPlotTools}
\title{UpSet Plot shows intersections by DE indications}
\usage{
upSetPlotTools(
  df,
  condition = "condition_default",
  pathOut = ".",
  writeData = TRUE
)
}
\arguments{
\item{df}{data.frame: output by 'listDeByTool' function.}

\item{condition}{string: name of analysed experminet (default: "condition_default")}

\item{pathOut}{string: path to write a pdf file with upset Plot (default: ".")}

\item{writeData}{boolean: TRUE if want write a PDF file (default: TRUE)}
}
\value{
data.frame with row sum by input df
}
\description{
UpSet Plot shows intersections by DE indications
}
\examples{
set.seed(42)
counts <- matrix(
  as.integer(c(
    rnbinom(200, mu = 10,  size = 1), 
    rnbinom(200, mu = 100, size = 1) 
  )),
  nrow = 100,
  dimnames = list(paste0("gene", seq_len(100)),
                    paste0("sample", seq_len(4)))
)
groups_info <- c("control", "control", "treated", "treated")
treats <- c("control", "treated")
cons_result <- runExpression(2, treats,rDataFrameCount = counts,
                            sepCharacter = ",",experimentName = "test_cons",
                            outDirPath = "." )
expDef_result <- expressionDefinition(resultTool = cons_result, groups = treats)
deByTool <- listDeByTool(consexpressionList = cons_result,
                        geneNames = rownames(counts),
                        deList = expDef_result)
if (requireNamespace("UpSetR", quietly = TRUE)) {
    upSetPlotData <- upSetPlotTools(df = deByTool,condition = "Control_vs_Treat",
                                pathOut = ".", writeData = FALSE)
}
}
