% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vennDiagramDe.R
\name{vennDiagramDe}
\alias{vennDiagramDe}
\title{Make a venn diagram with DE genes by consexpression2 results}
\usage{
vennDiagramDe(deList, outPath)
}
\arguments{
\item{deList}{Dataset eith consexpression2 results}

\item{outPath}{Path to write image}
}
\value{
void
}
\description{
Make a venn diagram with DE genes by consexpression2 results
}
