% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeResults.R
\name{writeResults}
\alias{writeResults}
\title{Write a table or data.frame in defined output file}
\usage{
writeResults(data, toolName = "toolDE_x", sepCharacter = "\\t")
}
\arguments{
\item{data}{table or data.frame dataset}

\item{toolName}{text with file name}

\item{sepCharacter}{character to separe data columns}
}
\value{
void
}
\description{
Write a table or data.frame in defined output file
}
