% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundanceKallistoImport.R
\name{abundanceKallistoImport}
\alias{abundanceKallistoImport}
\title{Import kallisto abundance to DESeq2 data set}
\usage{
abundanceKallistoImport(dir, pathReportRuns, pathReportFile, replics, dirRuns)
}
\arguments{
\item{dir}{path to kallisto results}

\item{pathReportRuns}{report from experiment contains columns pop : population,center: local from data,run: name for each run,condition: condition by run (ex.: treatment, treated))}

\item{pathReportFile}{parent directory that contains kallisto results organized by directory named like run}

\item{replics}{number of biological or technical replicates}

\item{dirRuns}{directory that contains kallisto results organized by directory named like run}
}
\value{
DESeq2 DataSet
}
\description{
Import kallisto abundance to DESeq2 data set
}
