% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensusList.R
\name{consensusList}
\alias{consensusList}
\title{consensusList: make a list of genes was consider DE by threshold tools}
\usage{
consensusList(consexpressionList, deTool, threshold = 5)
}
\arguments{
\item{consexpressionList}{list. Output by 'consexpressionR' function}

\item{deTool}{data.frame. Output of 'deByTool' function}

\item{threshold}{integer. Number of tool was consider a gene as DE to filter (default: 5)}
}
\value{
data.frame with DE genes and values analysis
}
\description{
consensusList: make a list of genes was consider DE by threshold tools
}
\examples{
data(gse95077)
treats <- c("BM", "JJ")
cons_result <- runExpression(numberReplics = 3,groupName = treats,
                            rDataFrameCount = gse95077,
                            sepCharacter = ",",
                            experimentName = "test_cons",
                            controlDeseq2 = "BM",
                            contrastDeseq2 = "JJ",
                            outDirPath = ".")
expDef_result <- expressionDefinition(resultTool = cons_result,
                                      groups = treats)
deByTool <- listDeByTool(consexpressionList = cons_result,
                         geneNames = row.names(gse95077),
                         deList = expDef_result)
consList <- consensusList(cons_result,deByTool)
}
