% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensusList.R
\name{consensusList}
\alias{consensusList}
\title{consensusList: make a list of genes was consider DE by threshold tools}
\usage{
consensusList(consexpressionList, deTool, threshold = 5)
}
\arguments{
\item{consexpressionList}{list. Output by 'consexpressionR' function}

\item{deTool}{data.frame. Output of 'deByTool' function}

\item{threshold}{integer. Number of tool was consider a gene as DE to filter (default: 5)}
}
\value{
data.frame with DE genes and values analysis
}
\description{
consensusList: make a list of genes was consider DE by threshold tools
}
\examples{
cons_result <- runExpression(numberReplics = 3, groupName = c("BM", "JJ"),
                              rDataFrameCount = gse95077,
                              sepCharacter = ",",
                              experimentName = "test_cons",
                              outDirPath = "." )
expDef_result <- expressionDefinition(resultTool = cons_result)
m <- as.matrix(cons_result)
deByTool <- listDeByTool(cons_result, row.names(m),expDef_result)
consList <- consensusList(cons_result,deByTool)
}
