% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consexpressionR.R
\name{consexpressionR}
\alias{consexpressionR}
\title{Main function. Make expression analysis of 7 tools and print output by consenso}
\usage{
consexpressionR(
  numberReplics,
  groupName,
  tableCountPath,
  sepCharacter = ",",
  rDataFrameCount = NULL,
  experimentName = "genericExperiment",
  outDirPath = "consexpression2_results/",
  methodNorm = "TMM",
  methodAdjPvalue = "BH",
  numberTopTable = 1e+06,
  printResults = FALSE,
  kallistoReport = "report.txt",
  kallistoDir = "kallisto_quant",
  kallistoSubDir = "expermient_kallisto",
  kallistoOut = "abundance.tsv",
  baySeqRespType = "Multiclass",
  knowSeqFilter = "ensembl_gene_id"
)
}
\arguments{
\item{numberReplics}{number of replicate (technical or biological) by sample}

\item{groupName}{text, name of samples or treatment}

\item{tableCountPath}{path to file csv that contains cout data or abundance data (local)}

\item{sepCharacter}{character used to split csv data, can be comma or tab}

\item{rDataFrameCount}{RData object with table count, where line name is gene name, and name column is treat name}

\item{experimentName}{text, name of experiment}

\item{outDirPath}{path to write output, need be a directory (local)}

\item{methodNorm}{normalization method to be used (limma)}

\item{methodAdjPvalue}{correction method, a character string. Can be abbreviated (limma)}

\item{numberTopTable}{maximum number of genes to list (limma)}

\item{printResults}{logical variable: TRUE print report by each tool, FALSE print only consensus result}

\item{kallistoReport}{path to kallisto report file _quant}

\item{kallistoDir}{name of kallistto results}

\item{kallistoSubDir}{name of directory in kallisto results}

\item{kallistoOut}{path to kallisto .tsv file}
}
\value{
list with all analisys expression
}
\description{
Main function. Make expression analysis of 7 tools and print output by consenso
}
\examples{
cons_result <- consexpressionR(numberReplics = 3, groupName = c("BM", "JJ"),
                              rDataFrameCount = table_count_df,
                              sepCharacter = ",",
                              experimentName = "test_cons",
                              outDirPath = "." )
}
