% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-ExpressionResultSet.R
\name{createExpressionResultSet}
\alias{createExpressionResultSet}
\title{Constructor for ExpressionResultSet}
\usage{
createExpressionResultSet(results, methods, parameters = list())
}
\arguments{
\item{results}{Named list of data.frames}

\item{methods}{Character vector with method names}

\item{parameters}{List of analysis parameters}
}
\value{
An ExpressionResultSet object
}
\description{
Constructor for ExpressionResultSet
}
