% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressionDefinition.R
\name{expressionDefinition}
\alias{expressionDefinition}
\title{Identify DE genes by method-specific thresholds}
\usage{
expressionDefinition(
  resultTool,
  groups = c(""),
  lfcMinLimma = -2,
  lfcMaxLimma = 2,
  pValueLimma = 0.05,
  FLimma = 0.8,
  lfcMinSamseq = -2,
  lfcMaxSamseq = 2,
  qValueSamseq = 0.05,
  scoreDSamseq = 0.8,
  lfcMinDeseq2 = -2,
  lfcMaxDeseq2 = 2,
  pValueDeseq2 = 0.05,
  lfcMinEdger = -2,
  lfcMaxEdger = 2,
  pValueEdger = 0.05,
  probNoiseq = 0.8,
  lfcMaxKnowseq = 2,
  lfcMinKnowseq = -2,
  pValueKnowseq = 0.05,
  deClassEbseq = "DE",
  ppThresholdEbseq = 0.8,
  printResults = FALSE,
  pathOutput = "."
)
}
\arguments{
\item{resultTool}{An \code{ExpressionResultSet} object.}

\item{groups}{Character vector of group names or conditions.}

\item{lfcMinLimma, ..., ppThresholdEbseq}{Numeric thresholds used per method (see details).}

\item{printResults}{Logical. Whether to write results to disk.}

\item{pathOutput}{Character. Directory path to save result tables.}
}
\value{
A list of data.frames with DE genes per method.
}
\description{
Applies specific filtering rules to each expression method result in the
ExpressionResultSet object and returns the DE genes per tool.
}
\examples{
data(gse95077)
treats <- c("BM", "JJ")
res <- runExpression(numberReplics = 3, groupName = treats, rDataFrameCount = gse95077, controlDeseq2 = "BM", contrastDeseq2 = "JJ" )
expDef <- expressionDefinition(resultTool = res, groups = treats)
}
