% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressionDefinition.R
\name{expressionDefinition}
\alias{expressionDefinition}
\title{Define who genes are differentially expressed by method}
\usage{
expressionDefinition(deTool, lfc)
}
\arguments{
\item{deTool}{table results by consexpression2 function}

\item{lfc}{minimum value to consider of log Fold Change}
}
\value{
void
}
\description{
Define who genes are differentially expressed by method
}
