% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressionDefinition.R
\name{expressionDefinition}
\alias{expressionDefinition}
\title{Define who genes are differentially expressed by method}
\usage{
expressionDefinition(
  resultTool,
  lfcMin = -2,
  lfcMax = 2,
  pValue = 0.05,
  prob = 0.95,
  qValue = 1,
  deClass = "DE"
)
}
\arguments{
\item{lfcMin}{minimum value to consider of log Fold Change (default: -2). Used by: KnowSeq, edgeR, limma, DESeq2 and SAMSeq.}

\item{lfcMax}{maximum value to consider of log Fold Change (default: 2), Used by: KnowSeq, edgeR, limma, DESeq2 and SAMSeq.}

\item{pValue}{maximum P-Value to consider (default: 0.05). Used by KnowSeq,, edgeR, limma and DESeq2, are consider p-Value >=.}

\item{prob}{floating point, minimum probability that a read count comes from a real gene expression peak, rather than background noiseq (default: 0.95)}

\item{qValue}{q-value is a measure of the statistical significance of the difference in expression between the compared groups, taking the problem of multiple comparisons into account (default: 1)}

\item{deClass}{name od class to consider by EBSeq (default: "DE")}

\item{deTool}{table results by consexpressionR function}
}
\value{
Object data.frame containing genes in the rows and tools executed in the columns, with a value of 1 for tools that considered the gene to be differentially expressed, and 0 for no DE.
}
\description{
Define who genes are differentially expressed by method
}
\examples{
cons_result <- consexpressionR(numberReplics = 3, groupName = c("BM", "JJ"),
                              rDataFrameCount = table_count_df,
                              sepCharacter = ",",
                              experimentName = "test_cons",
                              outDirPath = "." )
expDef_result <- expressionDefinition(resultTool = cons_result)
}
