% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frameAllGenes.R
\name{frameAllGenes}
\alias{frameAllGenes}
\title{frameAllGenes}
\usage{
frameAllGenes(cons_list, countMatrix)
}
\arguments{
\item{cons_list}{list generated by runExpression() function execution}

\item{countMatrix}{RData object with table count, where line name is gene name, and name column is treat name}
}
\value{
list with analysis expression containing: all methods executed and their results, ensuring thar all genes has been listed;
}
\description{
This function creates a data.frame for each result of the diffrential expression method performed. It was created to
ensure that all genes been listed, even thought the gene does not appear as diffrentially expressed, in this case NA data is inserted.
This functis is executed into runExpression() function, so there is no need to call it separatelly.
}
\examples{
data(gse95077)
treats = c("BM", "JJ")
cons_result <- runExpression(numberReplics = 3,
                             groupName = treats,
                              rDataFrameCount = gse95077,
                              sepCharacter = ",",
                              experimentName = "test_cons",
                              controlDeseq2 = "BM",
                              contrastDeseq2 = "JJ",
                              outDirPath = "." )
cons_rseult2 <- frameAllGenes(cons_list = cons_result,
                              countMatrix=gse95077)
}
