% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listDeByTool.R
\name{listDeByTool}
\alias{listDeByTool}
\title{listDeByTool: creates a data.frame with genes (rows) and DE tools (columns).
Each tool that identifies a gene as DE is marked with 1.}
\usage{
listDeByTool(consexpressionList, geneNames, deList)
}
\arguments{
\item{consexpressionList}{An \code{ExpressionResultSet} object with DE results.}

\item{geneNames}{Character vector. Names of all genes in the experiment.}

\item{deList}{List. Each element contains a character vector with DE gene names (from each tool).}
}
\value{
A data.frame with genes as rows and tools as columns; 1 indicates DE by that tool.
}
\description{
listDeByTool: creates a data.frame with genes (rows) and DE tools (columns).
Each tool that identifies a gene as DE is marked with 1.
}
\examples{
data(gse95077)
treats <- c("BM", "JJ")
cons_result <- runExpression(numberReplics = 3, groupName = treats, rDataFrameCount = gse95077, controlDeseq2 = "BM", contrastDeseq2 = "JJ" )
expDef_result <- expressionDefinition(resultTool = cons_result, groups = treats)
deByTool <- listDeByTool(cons_result, geneNames = rownames(gse95077), deList = expDef_result)

#      DESeq2 limma edgeR
# gene1      1     1     1
# gene2      0     0     1
# gene3      1     0     1
# gene4      0     1     0
# gene5      0     0     0
# gene6      1     1     1
}
