% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listDeByTool.R
\name{listDeByTool}
\alias{listDeByTool}
\title{listDeByTool: makes a data.frame with genes (lines) and all executed DE tool.
Each tool that identify gene as DE, are marked by 1.}
\usage{
listDeByTool(consexpressionList, geneNames, deList)
}
\arguments{
\item{consexpressionList}{A list generated by consexpressionR function}

\item{geneNames}{name of all genes in experiment (list)}

\item{deList}{List with names of all genes analysed in experiment}
}
\value{
data.frame with genes (lines) and DE tools (column), where 1 shows a
DE indication by tool.
}
\description{
listDeByTool: makes a data.frame with genes (lines) and all executed DE tool.
Each tool that identify gene as DE, are marked by 1.
}
\examples{
cons_result <- runExpression(numberReplics = 3, groupName = c("BM", "JJ"),
                              rDataFrameCount = gse95077,
                              sepCharacter = ",",
                              experimentName = "test_cons",
                              outDirPath = "." )
expDef_result <- expressionDefinition(resultTool = cons_result)
m <- as.matrix(gse95077)
deByTool <- listDeByTool(consexpressionList = cons_result,
                         geneNames = row.names(m),
                         deList = expDef_result)
}
