% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBaySeq.R
\name{runBaySeq}
\alias{runBaySeq}
\title{Execute baySeq gene Expression analysis}
\usage{
runBaySeq(countMatrix, designExperiment, out)
}
\arguments{
\item{countMatrix}{either a matrix of raw (read) counts.}

\item{designExperiment}{replicate and treatment by samples}

\item{out}{path to write baySeq report in csv file format}
}
\value{
baySeq report in data Frame fromat
}
\description{
Execute baySeq gene Expression analysis
}
\examples{
bayseqResult<-runBaySeq(countMatrix, replicates, createNameFileOutput(outDirPath,experimentName,execName='baySeq'))
}
