% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBaySeq.R
\name{runBaySeq}
\alias{runBaySeq}
\title{Execute baySeq gene Expression analysis}
\usage{
runBaySeq(
  countMatrix,
  groupName,
  numberReplics,
  sampleSize = 1000,
  estimationMethod = "QL",
  clusters = 2,
  equalDispersion = TRUE,
  priorProbabilities = c(0.5, 0.5),
  reestimationType = "BIC",
  topCountGroup = "DE",
  numberTopCount = 65000
)
}
\arguments{
\item{countMatrix}{either a matrix of raw (read) counts.}

\item{designExperiment}{replicate and treatment by samples}

\item{out}{path to write baySeq report in csv file format}
}
\value{
baySeq report in data Frame fromat
}
\description{
Execute baySeq gene Expression analysis
}
\examples{
bayseqResult<-runBaySeq(countMatrix, replicates, createNameFileOutput(outDirPath,experimentName,execName='baySeq'))
}
