% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBaySeq.R
\name{runBaySeq}
\alias{runBaySeq}
\title{Bayseq execution pre-definition}
\usage{
runBaySeq(
  countMatrix,
  groupName,
  numberReplics,
  sampleSize = 1000,
  estimationMethod = "QL",
  clusters = 2,
  equalDispersion = TRUE,
  priorProbabilities = c(0.5, 0.5),
  reestimationType = "BIC",
  topCountGroup = "DE",
  numberTopCount = 65000
)
}
\arguments{
\item{countMatrix}{table with mapped count}

\item{groupName}{string vector with names of treatment groups}

\item{numberReplics}{number of replics for each treatment}

\item{sampleSize}{length of sample to calculate probabilities}

\item{estimationMethod}{Defaults to "QL", indicating quasi-likelihood
estimation of priors. Currently, the only other possibilities are "ML", a
maximum-likelihood method, and "edgeR", the moderated dispersion estimates
produced by the 'edgeR' package.}

\item{clusters}{The number of nodes to be forked, if use parallel package}

\item{equalDispersion}{Should we assume equal dispersions of data across all
groups in the 'cD' object? Defaults to TRUE;}

\item{priorProbabilities}{value used by parameter ps by bayseq::getLikelihoods function}

\item{reestimationType}{value used by parameter pET by bayseq::getLikelihoods function}

\item{topCountGroup}{value used by parameter group by bayseq::topCounts function}

\item{numberTopCount}{value used by parameter number by bayseq::topCounts function}
}
\value{
extracted results from baySeq (generated by bayseq::topCounts function)
}
\description{
Bayseq execution pre-definition
}
