% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDeseq2.R
\name{runDeseq2}
\alias{runDeseq2}
\title{Execute DESeq2 DGE default analysis
is configured to read only kallisto files, and count data}
\usage{
runDeseq2(
  countMatrix,
  groupName,
  numberReplics,
  designExperiment,
  fitTypeParam = "local",
  pathDirRuns = ".",
  pathReportFile = "report_txi.txt",
  subDirRuns = "dir_runs",
  fileKallisto = "abundance.tsv"
)
}
\arguments{
\item{countMatrix}{original table count or abundance data}

\item{designExperiment}{replicate and treatment by samples}

\item{fitTypeParam}{either "parametric", "local", "mean", or "glmGamPoi" for the type of fitting of dispersions to the mean intensity}

\item{pathDirRuns}{parent directory that contains one folder with all kallisto output}

\item{pathReportFile}{file path with report by sample, conditions and runs}

\item{subDirRuns}{sub dir by pathDirRun}

\item{fileKallisto}{file output kallisto, exemaple: "abundance.tsv" or "quant.sf.gz"}
}
\value{
extracted results from DESeq2 (generated by DESeq2::results function)
}
\description{
Execute DESeq2 DGE default analysis
is configured to read only kallisto files, and count data
}
\examples{
df <- readCountFile(tableCountPath="data/table_count_df.csv",
                    split=",")
groupNameModel = c("BM","JJ")
numberReplicsModel = 3
designExperimentModel <- rep(groupNameModel, each = numberReplicModel)
toolResult <- NULL
toolResult$deseq2 <- runDeseq2(countMatrix = df,
                               groupName = groupNameModel,
                               numberReplics = numberReplicsModel)
}
