% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDeseq2.R
\name{runDeseq2}
\alias{runDeseq2}
\title{Execute DESeq2 DGE default analysis
is configured to read only kallisto files, and count data}
\usage{
runDeseq2(
  countMatrix,
  groupName,
  numberReplics,
  designExperiment,
  controlGroup = "",
  fitTypeParam = "local"
)
}
\arguments{
\item{countMatrix}{original table count or abundance data}

\item{groupName}{text, name of samples or treatment}

\item{numberReplics}{number of replicate (technical or biological) by sample}

\item{designExperiment}{replic and treatment by samples}

\item{controlGroup}{specific reference level}

\item{fitTypeParam}{either "parametric", "local", "mean", or "glmGamPoi" for the type of fitting of dispersions to the mean intensity}
}
\value{
extracted results from DESeq2 (generated by DESeq2::results function)
}
\description{
Execute DESeq2 DGE default analysis
is configured to read only kallisto files, and count data
}
\examples{
groupNameModel = c("BM","JJ")
numberReplicsModel = 3
toolResult <- NULL
m <- as.matrix(gse95077)
toolResult$deseq2 <- runDeseq2(countMatrix = m,
                               groupName = groupNameModel,
                               numberReplics = numberReplicsModel,
                               fitTypeParam = "local")
}
