% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBSeq.R
\name{runEbseq}
\alias{runEbseq}
\title{Execute EBSeq gene Expression analysis}
\usage{
runEbseq(
  countMatrix,
  designExperiment,
  fdr = 0.05,
  outFile = "consexpression2_noiseq.csv",
  sepCharacter
)
}
\arguments{
\item{countMatrix}{either a matrix of raw (read) counts.}

\item{designExperiment}{replicate and treatment by samples}

\item{outFile}{path to write EBSeq report in csv file format}

\item{sepCharacter}{}
}
\value{
EBSeq report in data Frame fromat
}
\description{
Execute EBSeq gene Expression analysis
}
