% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runEbseq.R
\name{runEbseq}
\alias{runEbseq}
\title{Execute EBSeq gene Expression analysis}
\usage{
runEbseq(countMatrix, designExperiment, fdr = 0.05, sepCharacter)
}
\arguments{
\item{countMatrix}{either a matrix of raw (read) counts.}

\item{designExperiment}{replicate and treatment by samples}

\item{fdr}{False Discovery Rate cutt off}

\item{sepCharacter}{pattern to seprare file csv}
}
\value{
EBSeq report in data Frame fromat
}
\description{
Execute EBSeq gene Expression analysis
}
