% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runEdger.R
\name{runEdger}
\alias{runEdger}
\title{Execute edgeR expression Analisys}
\usage{
runEdger(countMatrix, numberReplics, desingExperiment, edgerOutPath)
}
\arguments{
\item{countMatrix}{either a matrix of raw (read) counts.}

\item{numberReplics}{number of replicate (technical or biologcal) integer}

\item{desingExperiment}{replicate and treatment by samples}

\item{edgerOutPath}{path to write edgeR report in csv file format}
}
\value{
edgeR report in data Frame
}
\description{
Execute edgeR expression Analisys
}
