% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runEdger.R
\name{runEdger}
\alias{runEdger}
\title{Execute edgeR expression Analisys}
\usage{
runEdger(countMatrix, numberReplics, desingExperiment, methNorm = "TMM")
}
\arguments{
\item{countMatrix}{either a matrix of raw (read) counts.}

\item{numberReplics}{number of replicate (technical or biologcal) integer}

\item{desingExperiment}{replicate and treatment by samples}

\item{methNorm}{normalization method to be used in edgeR::calcNormFactors(), default: "TMM"}
}
\value{
edgeR report in data Frame
}
\description{
Execute edgeR expression Analisys
}
\examples{
data(gse95077)
treats = c("BM", "JJ")
numberReplicsModel = 3
designExperimentModel <- rep(treats, each = numberReplicsModel)
toolResult <- NULL
toolResult$edger <- runEdger(countMatrix = gse95077,
                               numberReplics = numberReplicsModel,
                               desingExperiment = designExperimentModel )
}
