% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runKnowSeq.R
\name{runKnowSeq}
\alias{runKnowSeq}
\title{Execute knowSeq differential Expression analyse}
\usage{
runKnowSeq(
  count,
  groupName,
  numberReplic,
  filterId = "ensembl_gene_id",
  notSapiens = FALSE
)
}
\arguments{
\item{count}{original table count or abundance data}

\item{groupName}{replicate and treatment by samples}

\item{numberReplic}{number of replicate (technical or biological) by sample}

\item{filterId}{The attribute used as filter to return the rest of the attributes.}

\item{notSapiens}{A boolean value that indicates if the user wants the human annotation or another annotation available in BiomaRt. The possible not human dataset can be consulted by calling the following function: biomaRt::listDatasets(useMart("ensembl")).}
}
\value{
extracted results from KnowSeq (generated by KnowSeq::DEGsExtraction() function)
}
\description{
Execute knowSeq differential Expression analyse
}
\examples{
\dontrun{

data(gse95077)
treats = c("BM", "JJ")
toolResult <- NULL
toolResult$knowSeq <- runKnowSeq(count = gse95077,  groupName = treats, numberReplic = 3)
 }
}
