% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runNoiSeq.R
\name{runNoiSeq}
\alias{runNoiSeq}
\title{Execute NOISeq gene Expression analysis}
\usage{
runNoiSeq(
  countMatrix,
  designExperiment,
  groups = c(""),
  normParm = "rpkm",
  kParam = 0.5,
  factorParam = c("Tissue"),
  lcParam = 0,
  replicatesParam = "technical",
  condExp = c("")
)
}
\arguments{
\item{countMatrix}{either a matrix of raw (read) counts.}

\item{designExperiment}{replicate and treatment by samples.}

\item{groups}{text list separated by comma, name of samples or treatment.}

\item{normParm}{Normalization method t can be one of "rpkm" (default), "uqua" (upper quartile), "tmm" (trimmed mean of M) or "n" (no normalization).}

\item{kParam}{Counts equal to 0 are replaced by k. By default, k = 0.5}

\item{factorParam}{A string indicating the name of factor whose levels are the conditions to be compared.}

\item{lcParam}{Length correction is done by dividing expression by length^lc. By default, lc = 0}

\item{replicatesParam}{In this argument, the type of replicates to be used is defined: "technical", "biological" or "no" replicates. By default, "technical" replicates option is chosen.}

\item{condExp}{A vector containing the two conditions to be compared by the differential expression algorithm (needed when the factor contains more than 2 different conditions).}
}
\value{
A data.frame with output of noiseqOut@results
}
\description{
Execute NOISeq gene Expression analysis
}
\examples{
data(gse95077)
treats = c("BM", "JJ")
toolResult <- NULL
toolResult$noiseq <- runNoiSeq(countMatrix = gse95077,
                               designExperiment = rep(treats, each = 3))
}
