% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runNoiSeq.R
\name{runNoiSeq}
\alias{runNoiSeq}
\title{Execute NOISeq gene Expression analysis}
\usage{
runNoiSeq(countMatrix, designExperiment)
}
\arguments{
\item{countMatrix}{either a matrix of raw (read) counts.}

\item{designExperiment}{replicate and treatment by samples}

\item{noiseqOutPtah}{path to write EBSeq report in csv file format}
}
\value{
NOISeq report in data Frame fromat
}
\description{
Execute NOISeq gene Expression analysis
}
\examples{
noiseqResult<-runNOISeq(countMatrix, designExperiment, createNameFileOutput(outDirPath,experimentName,execName='NOISeq'))
}
