% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSamSeq.R
\name{runSamSeq}
\alias{runSamSeq}
\title{Execute SAMSeq gene Expression analysis to count data}
\usage{
runSamSeq(
  countMatrix,
  designExperiment,
  respType = "Two class unpaired",
  numberPermutations = 100
)
}
\arguments{
\item{countMatrix}{either a matrix of raw (read) counts.}

\item{replic}{number of replicate (technical or biologcal) integer}

\item{desingExperiment}{replicate and treatment by samples}

\item{samseqOutPath}{path to write SAMSeq report in csv file format}
}
\value{
SAMSeq report in data Frame
}
\description{
Execute SAMSeq gene Expression analysis to count data
}
