% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-ExpressionResultSet.R
\name{show,ExpressionResultSet-method}
\alias{show,ExpressionResultSet-method}
\title{Show method for ExpressionResultSet}
\usage{
\S4method{show}{ExpressionResultSet}(object)
}
\arguments{
\item{object}{An ExpressionResultSet object}
}
\description{
Show method for ExpressionResultSet
}
