% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-ExpressionResultSet.R
\name{summary,ExpressionResultSet-method}
\alias{summary,ExpressionResultSet-method}
\title{Summarize an ExpressionResultSet object}
\usage{
\S4method{summary}{ExpressionResultSet}(object)
}
\arguments{
\item{object}{An ExpressionResultSet object.}
}
\value{
Invisibly returns NULL.
}
\description{
Summarize an ExpressionResultSet object
}
