% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upSetPlotTools.R
\name{upSetPlotTools}
\alias{upSetPlotTools}
\title{UpSet Plot shows intersections by DE indications}
\usage{
upSetPlotTools(
  df,
  condition = "condition_default",
  pathOut = ".",
  writeData = TRUE
)
}
\arguments{
\item{df}{data.frame: output by 'listDeByTool' function.}

\item{condition}{string: name of analysed experminet (default: "condition_default")}

\item{pathOut}{string: path to write a pdf file with upset Plot (default: ".")}

\item{writeData}{boolean: TRUE if want write a PDF file (default: TRUE)}
}
\value{
data.frame with row sum by input df
}
\description{
UpSet Plot shows intersections by DE indications
}
\examples{
cons_result <- consexpressionR(numberReplics = 3, groupName = c("BM", "JJ"),
                              rDataFrameCount = table_count_df,
                              sepCharacter = ",",
                              experimentName = "test_cons",
                              outDirPath = "." )
expDef_result <- expressionDefinition(resultTool = cons_result)
deByTool <- listDeByTool(cons_result, expDef_result)
upSetPlotData <- upSetPlotTools(df = deByTool,
                                condition = "Control_vs_Treat",
                                pathOut = ".",
                                writeData = TRUE)
}
